Move prescription comparison to separate file to match test name
