+++ coordinator/src/services/verification/signature-verification.ts
@@ -1,26 +1,10 @@
 import {ElementCompact} from "xml-js"
-import {common, hl7V3} from "@models"
+import {hl7V3} from "@models"
 import {writeXmlStringCanonicalized} from "../serialisation/xml"
 import {convertFragmentsToHashableFormat, extractFragments} from "../translation/request/signature"
 import {createParametersDigest} from "../translation/request"
 import crypto from "crypto"
 import {isTruthy} from "../translation/common"
-
-function comparePrescriptions(p1: common.Prescription, p2: common.Prescription): Array<string> {
-  // TODO: AEA-2645 + AEA-2524 - Add key fields to be compared
-  const p1KeyValues = Object.entries(p1)
-  const p2KeyValues = Object.entries(p2)
-  return p1KeyValues.map((keyValue, index) => {
-    if (keyValue[1] !== p2KeyValues[index][1]) {
-      const camelCaseName = `${keyValue[0]}`
-      const firstLetterUpperCase = camelCaseName.substring(0, 1).toUpperCase()
-      const allOtherLetters = camelCaseName.substring(1)
-      const pascalCaseName = `${firstLetterUpperCase}${allOtherLetters}`
-      const titleCaseName = pascalCaseName.replace(/([A-Z])/g, " $1").trim()
-      return `${titleCaseName} does not match`
-    }
-  }).filter(Boolean)
-}
 
 function verifySignature(parentPrescription: hl7V3.ParentPrescription): Array<string> {
   const validSignatureFormat = verifySignatureHasCorrectFormat(parentPrescription)
@@ -113,7 +97,6 @@ function verifyCertificate(parentPrescription: hl7V3.ParentPrescription): boolea
 }
 
 export {
-  comparePrescriptions,
   extractSignatureRootFromParentPrescription,
   verifySignatureDigestMatchesPrescription,
   verifyPrescriptionSignatureValid,
